fix: exclude abi-dependents from tinygo build
